fix: float4 precision loss and sslmode=prefer incorrectly requiring SSL
- float4-type.ts: parseBinary rounded every decoded value to 2 decimal
  places (Math.round(...*100)/100), silently truncating float4 precision
  on every query. Confirmed against PostgreSQL's float4send/float4recv
  and pq_sendfloat4/pq_getmsgfloat4 (raw IEEE754, no rounding) and
  node-pg-types' parseFloat32 (plain readFloatBE). Now matches float8-type.ts.
  Updated float4.spec.ts and 04-query.spec.ts expectations to use
  Math.fround() for values not exactly representable in 32-bit float,
  since the prior rounding was masking float4's real precision.
- connection-config.ts: sslmode=prefer was mapped to requireSSL:true,
  making pg-socket.ts treat a server's SSL refusal as fatal. Per libpq
  semantics, prefer should attempt SSL and silently fall back to a plain
  connection when refused; only require/verify-ca/verify-full should
  hard-fail. Removed 'prefer' from the requireSSL list.

+++ test/B-connection/04-query.spec.ts
@@ -190,7 +190,9 @@ describe('query() (Extended Query)', () => {
     expect(row.f_int2).toStrictEqual(1);
     expect(row.f_int4).toStrictEqual(12345);
     expect(row.f_int8).toStrictEqual(BigInt('9007199254740995'));
-    expect(row.f_float4).toStrictEqual(1.2);
+    // float4 is IEEE754 single-precision; 1.2 is not exactly representable
+    // in 32 bits, so the stored value widens to the nearest float4 value.
+    expect(row.f_float4).toStrictEqual(Math.fround(1.2));
     expect(row.f_float8).toStrictEqual(5.12345);
     expect(row.f_char).toStrictEqual('a');
     expect(row.f_varchar).toStrictEqual('abcd');
